ACE removal - phase 2/2.
ACE is now completely removed from the project. The remaining code is now using the new threading and networking system.

+++ src/game/WorldHandlers/MiscHandler.cpp
@@ -836,7 +836,7 @@ void WorldSession::HandleAreaTriggerOpcode(WorldPacket& recv_data)
         }
 
         // need find areatrigger to inner dungeon for landing point
-        if (at->target_mapId != corpseMapId)
+        if (at->target_mapId != uint32(corpseMapId))
         {
             if (AreaTrigger const* corpseAt = sObjectMgr.GetMapEntranceTrigger(corpseMapId))
             {
